requirements.txt
python-telegram-bot[webhooks]
psycopg2
requests
bs4
geopy
beautifulsoup4
playwright
uvicorn
starlette
googlemaps